chore: organized components per page
