+++ frontend/src/components/homepage/index.ts
@@ -1,9 +1,4 @@
-export * from "./pagination";
-export * from "./production-list";
 export * from "./masthead";
 export * from "./search-bar";
 export * from "./featured-section";
-export * from "./archive-sidebar";
 export * from "./search-header";
-export * from "./search-hero";
-export * from "./results-bar";
