+++ frontend/src/components/searchpage/index.ts
@@ -1,0 +1,5 @@
+export * from "./pagination";
+export * from "./production-list";
+export * from "./search-hero";
+export * from "./results-bar";
+export * from "./archive-sidebar";
