+++ frontend/src/components/searchpage/production-list/ProductionList.tsx
