+++ frontend/src/components/searchpage/production-list/index.ts
