+++ frontend/src/components/searchpage/results-bar/ResultsBar.tsx
