+++ frontend/src/components/searchpage/results-bar/index.ts
