+++ frontend/src/components/searchpage/search-hero/SearchHero.tsx
