+++ frontend/src/components/searchpage/search-hero/index.ts
